requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-google-genai
google-generativeai
chromadb
sentence-transformers
pypdf
python-dotenv
